+++ tomcat/src/main/java/org/apache/coyote/http11/processor/OutputProcessor.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.processor;
 
 import java.io.IOException;
 import java.io.OutputStream;
