+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -7,7 +7,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import org.apache.coyote.http11.HttpCookie;
-import org.apache.coyote.http11.RequestPathType;
 import org.apache.coyote.http11.request.requestLine.HttpRequestLine;
 import org.apache.coyote.http11.request.requestLine.MethodType;
