+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestPathType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 public enum RequestPathType {
